pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prompthunt-reddit-agent"
version = "0.1.0"
description = "Compliance-first Reddit opportunity agent MVP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aiosqlite>=0.21.0",
  "asyncpg>=0.30.0",
  "fastapi>=0.116.0",
  "greenlet>=3.2.4",
  "httpx>=0.28.1",
  "langgraph>=0.2.60",
  "pydantic-settings>=2.10.1",
  "python-dotenv>=1.1.1",
  "pyyaml>=6.0.2",
  "redis>=6.4.0",
  "sqlalchemy>=2.0.43",
  "temporalio>=1.17.0",
  "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.1",
  "pytest-asyncio>=1.1.0",
  "ruff>=0.12.10",
]

[tool.hatch.build.targets.wheel]
packages = ["src/reddit_agent"]

[tool.pytest.ini_options]
pythonpath = ["src", "."]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["legacy"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "single"
